This commit updates PAC so that we can build an image using PAC 2.0.X.
I've decided that we might as well version control the Dockerfile alongside PAC since bundler
can then be used to install the 'right' dependencies for the image. This also makes it much
easier to build and test the docker image together with the application itself.

I've also fixed it so that there is no need to pass the display parameters to
the container. That was fixed by using a much newer version of wkhtmltopdf. (~ no longer 4 years old.)